backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai==0.3.2
pydantic==2.5.0
Pillow==10.1.0
python-multipart==0.0.6